requirements.txt
django
djangorestframework
django-cors-headers
xarray
netcdf4
pandas